Update colormap arguments for consistency, can supply a colormap name or instance
