+++ ternary/heatmapping.py
@@ -2,87 +2,11 @@
 Various Heatmaps.
 """
 
-import matplotlib
-from matplotlib import pyplot
-from matplotlib.colors import rgb2hex
 import numpy
 
 from helpers import SQRT3, SQRT3OVER2, unzip, normalize, simplex_iterator
 import plotting
-
-
-## Default colormap, other options here: http://www.scipy.org/Cookbook/Matplotlib/Show_colormaps
-DEFAULT_COLOR_MAP_NAME = 'jet'
-
-## Matplotlib Colormapping ##
-
-def get_cmap(cmap_name=None):
-    """
-    Loads a matplotlib colormap if specified or supplies the default.
-
-    Parameters
-    ----------
-    cmap_name, string
-        The name of the Matplotlib colormap to look up.
-
-    Returns
-    -------
-    The desired Matplotlib colormap
-
-    Raises
-    ------
-    ValueError if colormap name is not recognized by Matplotlib
-    """
-
-    if not cmap_name:
-        cmap_name = DEFAULT_COLOR_MAP_NAME
-    return pyplot.get_cmap(cmap_name)
-
-def colormapper(value, lower=0, upper=1, cmap=None):
-    """
-    Maps values to colors by normalizing within [a,b], obtaining rgba from the
-    given matplotlib color map for heatmap polygon coloring.
-
-    Parameters
-    ----------
-    x: float
-        The value to be colormapped
-    a: float
-        Lower bound of colors
-    b: float
-        Upper bound of colors
-    cmap, Matplotlib Colormap Object (optional)
-        Colormap object to prevent repeated lookup
-
-    Returns
-    -------
-    hex_, float
-        The value mapped to an appropriate RGBA color value
-    """
-
-    if not cmap:
-        cmap = get_cmap()
-    if upper - lower == 0:
-        rgba = cmap(0)
-    else:
-        rgba = cmap((value - lower) / float(upper - lower))
-    hex_ = rgb2hex(rgba)
-    return hex_
-
-def colorbar_hack(ax, vmin, vmax, cmap, scientific=False):
-    """Colorbar hack to insert colorbar on ternary plot. Called by heatmap, 
-    not intended for direct usage."""
-    # http://stackoverflow.com/questions/8342549/matplotlib-add-colorbar-to-a-sequence-of-line-plots
-    norm = pyplot.Normalize(vmin=vmin, vmax=vmax)
-    sm = pyplot.cm.ScalarMappable(cmap=cmap, norm=norm)
-    # Fake up the array of the scalar mappable. Urgh...
-    sm._A = []
-    cb = pyplot.colorbar(sm, ax=ax, format='%.4f')
-    cb.locator = matplotlib.ticker.LinearLocator(numticks=7)
-    if scientific:
-        cb.formatter = matplotlib.ticker.ScalarFormatter()
-        cb.formatter.set_powerlimits((0, 0))
-    cb.update_ticks()
+from colormapping import get_cmap, colormapper, colorbar_hack
 
 ### Heatmap Triangulation Coordinates ###
 
@@ -190,7 +114,7 @@ def hexagon_coordinates(i, j, k):
 
 ## Heatmaps ##
 
-def heatmap(d, scale, vmin=None, vmax=None, cmap_name=None, ax=None,
+def heatmap(d, scale, vmin=None, vmax=None, cmap=None, ax=None,
             scientific=False, style='triangular', colorbar=True):
     """
     Plots heatmap of given color values.
@@ -206,7 +130,7 @@ def heatmap(d, scale, vmin=None, vmax=None, cmap_name=None, ax=None,
         The minimum color value, used to normalize colors. Computed if absent.
     vmax: float, None
         The maximum color value, used to normalize colors. Computed if absent.
-    cmap_name: String, None
+    cmap: String or matplotlib.colors.Colormap, None
         The name of the Matplotlib colormap to use.
     ax: Matplotlib AxesSubplot, None
         The subplot to draw on.
@@ -224,7 +148,7 @@ def heatmap(d, scale, vmin=None, vmax=None, cmap_name=None, ax=None,
     
     if not ax:
         fig, ax = pyplot.subplots()
-    cmap = get_cmap(cmap_name)
+    cmap = get_cmap(cmap)
     if not vmin:
         vmin = min(d.values())
     if not vmax:
@@ -255,7 +179,7 @@ def heatmap(d, scale, vmin=None, vmax=None, cmap_name=None, ax=None,
 
 ## User Convenience Functions ##
 
-def heatmapf(func, scale=10, boundary=True, cmap_name=None,
+def heatmapf(func, scale=10, boundary=True, cmap=None,
                         ax=None, scientific=False, style='triangular',
                         colorbar=True):
     """
@@ -269,7 +193,7 @@ def heatmapf(func, scale=10, boundary=True, cmap_name=None,
         The scale used to partition the simplex
     boundary: Bool, True
         Include the boundary points or not
-    cmap_name: String, None
+    cmap: String, None
         The name of the Matplotlib colormap to use
     ax: Matplotlib axis object, None
         The axis to draw the colormap on
@@ -290,7 +214,7 @@ def heatmapf(func, scale=10, boundary=True, cmap_name=None,
     for i, j, k in simplex_iterator(scale=scale, boundary=boundary):
         d[(i, j)] = func(normalize([i, j, k]))
     # Pass everything to the heatmapper
-    ax = heatmap(d, scale, cmap_name=cmap_name, ax=ax, style=style,
+    ax = heatmap(d, scale, cmap=cmap, ax=ax, style=style,
                            scientific=scientific, colorbar=colorbar)
     return ax
